checkout v0.1.0 Basket.Manager

Basket Manager module. Used for manipulate products in basket

Functions

add_item(item)

add_item(Struct.t()) :: Atom.t()

Update Agent state Steps:

Get Basket from Agent

Insert new product in Basket item list and set timestamp to item struct

Apply discount rules

recalculate Basket total and discount

save new state to Gen Server

Returns :ok.

apply_rules(b)

apply_rules(Basket.t()) :: List.t()

Iterate through rules and apply each rule to the product list. But before we should set all discount to zero

Returns List.

basket()

Get state from Agent. Returns %Basket{}.

basket(basket)

Set state to Agent. Returns :ok.

init(rule_list)

init(List.t()) :: Struct.t()

Starts Gen Server and hold basket as state of it Returns {:ok, #PID}.

recalculate(basket)

recalculate(Basket.t()) :: Basket.t()

Update Basket’s total and discount by calculating all items discount and items price separately

Returns %Busket{}.

reset(rule_list)

Set empty List in state of Agent.

Returns :ok.