chexkout v0.1.0 Rule.Discount protocol

A protocol for dealing with the various forms of rule resources. For now there are rules only for Vouchers and Tshirts. Protocol needs to be implemented by all rules

Types

t()

t() :: term()

Functions

calculate(rule)

calculate(Rule.t()) :: List.t()

All rules has to implement protocol’s method calculate